+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - \#3559 - Create Modal: always show Port Index for host ports
 - \#3424 - Maintain "assign random port" settings after JSON mode switch
 - \#3554 - Remove unncessary active class
+- \#3692 - Remove the Deployment Stop Button in the UI
 
 ### Added
 - VIP labels in port definitions